[upgrade_path] Move the upgrade_path function to upgrade_helpers script
